WIP - Server functions testing
