+++ apinf_packages/ratings/collection/server/publications.js
@@ -10,22 +10,15 @@ import { check } from 'meteor/check';
 // Collection imports
 import ApiBackendRatings from '/apinf_packages/ratings/collection';
 
+// Function imports
+import serverFunctions from './functions.js';
+
+// Deconstruct server serverFunctions
+const { publishMyApiBackendRating } = serverFunctions;
+console.log(publishMyApiBackendRating)
 // User rating for a single API Backend
-Meteor.publish('myApiBackendRating', function (apiBackendId) {
-  // Make sure apiBackendId is a String
-  check(apiBackendId, String);
-
-  // get current user ID
-  const userId = this.userId;
-
-  // get user API Backend rating
-  const userApiBackendRatings = ApiBackendRatings.find({
-    userId,
-    apiBackendId,
-  });
-
-  return userApiBackendRatings;
-});
+Meteor.publish('myApiBackendRating',
+  publishMyApiBackendRating({ check, ApiBackendRatings }));
 
 // User ratings for all API Backends
 Meteor.publish('myApiBackendRatings', function () {
